CREATE OR REPLACE ersetzt durch CREATE, da sonst jedes mal die Tabell…

+++ backend/src/main/resources/db/migration/V4__add_device_state_columns.sql
@@ -1,6 +1,6 @@
 ALTER TABLE devices
     ADD COLUMN state_on       BOOLEAN          NOT NULL DEFAULT FALSE,
     ADD COLUMN brightness     INT              NOT NULL DEFAULT 50,
-    ADD COLUMN temperature    DOUBLE           NOT NULL DEFAULT 21.0,
-    ADD COLUMN sensor_value   DOUBLE           NOT NULL DEFAULT 0,
+    ADD COLUMN temperature    DOUBLE PRECISION NOT NULL DEFAULT 21.0,
+    ADD COLUMN sensor_value   DOUBLE PRECISION NOT NULL DEFAULT 0,
     ADD COLUMN cover_position INT              NOT NULL DEFAULT 0;
